requirements.txt
streamlit==1.28.1
graphviz==0.20.1
pandas==2.0.0
numpy==1.24.0
pytest==7.0.0
fpdf2==2.7.0